Create ReadPathAggregationDemo
