+++ demo-dvf/src/main/java/io/mincong/dvf/demo/Main.java
@@ -3,6 +3,7 @@ package io.mincong.dvf.demo;
 public class Main {
 
   public static void main(String[] args) {
-    new WritePathDemo().run();
+    //    new WritePathDemo().run();
+    new ReadPathAggregationDemo().run();
   }
 }
